+++ plugins/woocommerce-admin/client/header/activity-panel/index.js
@@ -14,24 +14,31 @@ import {
 import { getAdminLink } from '@woocommerce/wc-admin-settings';
 import { H, Section } from '@woocommerce/components';
 import {
+	ONBOARDING_STORE_NAME,
 	OPTIONS_STORE_NAME,
 	useUser,
-	ONBOARDING_STORE_NAME,
 } from '@woocommerce/data';
 import { getHistory, getNewPath } from '@woocommerce/navigation';
 import { recordEvent } from '@woocommerce/tracks';
+import { applyFilters } from '@wordpress/hooks';
 
 /**
  * Internal dependencies
  */
 import './style.scss';
-import { getUnreadNotes } from './unread-indicators';
+import { isNotesPanelVisible } from './unread-indicators';
 import { isWCAdmin } from '../../dashboard/utils';
 import { Tabs } from './tabs';
 import { SetupProgress } from './setup-progress';
 import { DisplayOptions } from './display-options';
 import { HighlightTooltip } from './highlight-tooltip';
 import { Panel } from './panel';
+import {
+	getLowStockCount as getLowStockProducts,
+	getOrderStatuses,
+	getUnreadOrders,
+} from '../../homescreen/activity-panel/orders/utils';
+import { getUnapprovedReviews } from '../../homescreen/activity-panel/reviews/utils';
 
 const HelpPanel = lazy( () =>
 	import( /* webpackChunkName: "activity-panels-help" */ './panels/help' )
@@ -39,7 +46,7 @@ const HelpPanel = lazy( () =>
 
 const InboxPanel = lazy( () =>
 	import(
-		/* webpackChunkName: "activity-panels-inbox" */ '../../inbox-panel'
+		/* webpackChunkName: "activity-panels-inbox" */ './panels/inbox/inbox-panel'
 	)
 );
 
@@ -65,8 +72,51 @@ export const ActivityPanel = ( { isEmbedded, query, userPreferencesData } ) => {
 		return trackData;
 	};
 
+	function getThingsToDoNextCount(
+		tasks,
+		dismissedTasks,
+		isExtendedTaskListHidden
+	) {
+		if ( ! tasks || isExtendedTaskListHidden ) {
+			return 0;
+		}
+		return tasks.filter(
+			( task ) =>
+				task.visible &&
+				! task.completed &&
+				! dismissedTasks.includes( task.key )
+		).length;
+	}
+
+	function isAbbreviatedPanelVisible(
+		select,
+		setupTaskListHidden,
+		thingsToDoNextCount
+	) {
+		const orderStatuses = getOrderStatuses( select );
+
+		const isOrdersCardVisible = setupTaskListHidden
+			? getUnreadOrders( select, orderStatuses ) > 0
+			: false;
+		const isReviewsCardVisible = setupTaskListHidden
+			? getUnapprovedReviews( select )
+			: false;
+		const isLowStockCardVisible = setupTaskListHidden
+			? getLowStockProducts( select )
+			: false;
+
+		return (
+			thingsToDoNextCount > 0 ||
+			isOrdersCardVisible ||
+			isReviewsCardVisible ||
+			isLowStockCardVisible
+		);
+	}
+
 	const {
 		hasUnreadNotes,
+		hasAbbreviatedNotifications,
+		thingsToDoNextCount,
 		requestingTaskListOptions,
 		setupTaskListComplete,
 		setupTaskListHidden,
@@ -74,9 +124,31 @@ export const ActivityPanel = ( { isEmbedded, query, userPreferencesData } ) => {
 		previewSiteBtnTrackData,
 	} = useSelect( ( select ) => {
 		const { getOption, isResolving } = select( OPTIONS_STORE_NAME );
+		const isSetupTaskListHidden =
+			getOption( 'woocommerce_task_list_hidden' ) === 'yes';
+		const isExtendedTaskListHidden =
+			getOption( 'woocommerce_extended_task_list_hidden' ) === 'yes';
+		const extendedTaskList = applyFilters(
+			'woocommerce_admin_onboarding_task_list',
+			[],
+			query
+		);
+		const dismissedTasks =
+			getOption( 'woocommerce_task_list_dismissed_tasks' ) || [];
+		const thingsToDoCount = getThingsToDoNextCount(
+			extendedTaskList,
+			dismissedTasks,
+			isExtendedTaskListHidden
+		);
 
 		return {
-			hasUnreadNotes: getUnreadNotes( select ),
+			hasUnreadNotes: isNotesPanelVisible( select ),
+			hasAbbreviatedNotifications: isAbbreviatedPanelVisible(
+				select,
+				isSetupTaskListHidden,
+				thingsToDoCount
+			),
+			thingsToDoNextCount: thingsToDoCount,
 			requestingTaskListOptions:
 				isResolving( 'getOption', [
 					'woocommerce_task_list_complete',
@@ -84,8 +156,7 @@ export const ActivityPanel = ( { isEmbedded, query, userPreferencesData } ) => {
 				isResolving( 'getOption', [ 'woocommerce_task_list_hidden' ] ),
 			setupTaskListComplete:
 				getOption( 'woocommerce_task_list_complete' ) === 'yes',
-			setupTaskListHidden:
-				getOption( 'woocommerce_task_list_hidden' ) === 'yes',
+			setupTaskListHidden: isSetupTaskListHidden,
 			trackedCompletedTasks:
 				getOption( 'woocommerce_task_list_tracked_completed_tasks' ) ||
 				[],
@@ -147,7 +218,7 @@ export const ActivityPanel = ( { isEmbedded, query, userPreferencesData } ) => {
 			name: 'inbox',
 			title: __( 'Inbox', 'woocommerce-admin' ),
 			icon: <Icon icon={ inboxIcon } />,
-			unread: hasUnreadNotes,
+			unread: hasUnreadNotes || hasAbbreviatedNotifications,
 			visible:
 				( isEmbedded || ! isHomescreen() ) && ! isPerformingSetupTask(),
 		};
@@ -225,7 +296,14 @@ export const ActivityPanel = ( { isEmbedded, query, userPreferencesData } ) => {
 
 		switch ( tab ) {
 			case 'inbox':
-				return <InboxPanel />;
+				return (
+					<InboxPanel
+						hasAbbreviatedNotifications={
+							hasAbbreviatedNotifications
+						}
+						thingsToDoNextCount={ thingsToDoNextCount }
+					/>
+				);
 			case 'help':
 				return <HelpPanel taskName={ task } />;
 			default: